Added update movie funtionality
