+++ MovieApi/Controllers/MovieController.cs
@@ -3,6 +3,8 @@ using Microsoft.Extensions.Logging;
 using MovieApi.Data;
 using MovieApi.DTOs;
 using MovieApi.Services;
+using System;
+using System.Threading.Tasks;
 
 namespace MovieApi.Controllers
 {
@@ -51,5 +53,26 @@ namespace MovieApi.Controllers
             }
             return BadRequest();
         }
+        [HttpPut("{ID}")]
+        public async Task<IActionResult> UpdateMovie([FromBody] UpdateMovieDto model, string ID)
+        {
+            if (!ModelState.IsValid) return BadRequest();
+
+            try
+            {
+
+                var result = await _movieRepository.UpdateMovie(ID, model);
+
+
+                return Ok(new { result });
+            }
+            catch (Exception e)
+            {
+                _logger.LogError(e.Message);
+                return BadRequest(e.Message);
+
+            }
+        }
+
     }
 }
