+++ MovieApi/Models/Movie.cs
@@ -10,7 +10,7 @@ namespace MovieApi.Models
         public string MovieId { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public DateTime RealeaseDate { get; set; }
+        public DateTime ReleaseDate { get; set; }
         public int Rating { get; set; }
         public double TicketPrice { get; set; }
         public string Country { get; set; }
