+++ MovieApi/Services/IMovieRepository.cs
@@ -1,7 +1,4 @@
 ﻿using MovieApi.DTOs;
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 
 namespace MovieApi.Services
@@ -10,5 +7,6 @@ namespace MovieApi.Services
     {
         public string AddMovie(MovieDTO movie);
         public bool RemoveMovie(string Id);
+        Task<string> UpdateMovie(string Id, UpdateMovieDto model);
     }
 }
