+++ MovieApi/Services/MovieRepository.cs
@@ -1,4 +1,8 @@
-﻿using System;
+﻿using Microsoft.EntityFrameworkCore;
+using MovieApi.Data;
+using MovieApi.DTOs;
+using MovieApi.Models;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -12,6 +16,27 @@ namespace MovieApi.Services
         public MovieRepository(AppDbContext ctx)
         {
             _ctx = ctx;
+        }
+
+        public async Task<string> UpdateMovie(string Id, UpdateMovieDto model)
+        {
+            var movie = await _ctx.Movies.FirstOrDefaultAsync(x => x.MovieId == Id);
+
+            if (movie == null) return "failed to update movie";
+
+
+            movie.Country = !string.IsNullOrEmpty(model.Country) ? model.Country : movie.Country;
+            movie.Description = !string.IsNullOrEmpty(model.Description) ? model.Description : movie.Description;
+            movie.Rating = model.Rating > 0 ? model.Rating : movie.Rating;
+            movie.Name = !string.IsNullOrEmpty(model.Name) ? model.Name : movie.Name;
+            movie.TicketPrice = model.TicketPrice > 0 ? model.TicketPrice : movie.TicketPrice;
+            movie.PhotoUrl = !string.IsNullOrEmpty(model.PhotoUrl) ? model.PhotoUrl : movie.PhotoUrl;
+            movie.ReleaseDate = !model.ReleaseDate.Equals(movie.ReleaseDate) ? model.ReleaseDate : movie.ReleaseDate;
+
+            _ctx.Movies.Update(movie);
+            await _ctx.SaveChangesAsync();
+
+            return "Update done successfully";
         }
         //This method is what removes a movie from the database.
         public string AddMovie(MovieDTO movie)
@@ -34,8 +59,8 @@ namespace MovieApi.Services
                     MovieId = Guid.NewGuid().ToString(),// The guid method is used to create Id for the MovieId
                     Name = movie.Name,
                     Description = movie.Description,
-                    ReleaseDate = movie.ReleaseDate,
-                    Rating = movie.Rating,
+                    ReleaseDate = DateTime.Parse(movie.ReleaseDate),
+                    Rating = int.Parse(movie.Rating),
                     TicketPrice = movie.TicketPrice,
                     Country = movie.Country,
                     PhotoUrl = movie.PhotoUrl,
